+++ web/documentserver-example/ruby/app/models/document_helper.rb
@@ -149,8 +149,10 @@ class DocumentHelper
 
     # get the correct file name if such a name already exists
     def get_correct_name(file_name, user_address)
+      maxName = 50
       ext = File.extname(file_name)  # get file extension
-      base_name = File.basename(file_name, ext)  # get file name without extension
+      # get file name without extension
+      base_name = File.basename(file_name, ext)[0...maxName] + (file_name.length > maxName ? '[...]' : '')
       name = base_name + ext.downcase  # get full file name
       index = 1