+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 
 - python: user avatar
+- nodejs: user avatar
+- trimming long name of uploading file
+- nodejs: link in referenceData
 - onRequestSelectDocument method
 - onRequestSelectSpreadsheet method
 - key in referenceData